+++ src/logging/WriterFrontend.h
@@ -66,20 +66,6 @@ private:
     std::vector<LogRecord> records;
 };
 
-
-/**
- * A struct holding all ncessary information that relates to
- * log writes for a given path. These values are constant over
- * the lifetime of a \a WriterFrontend.
- */
-struct LogWriteHeader {
-    EnumValPtr stream_id;                 // The enum identifying the stream.
-    EnumValPtr writer_id;                 // The enum identifying the writer for backwards compat.
-    std::string filter_name;              // The name of the filter.
-    std::string path;                     // The path as configured or produced by the filter's path_func.
-    std::vector<threading::Field> fields; // The schema describing a log record.
-};
-
 } // namespace detail
 
 /**
